Member entity - userDetail : separation of concerns
